move progress bar to header and adjust style
