+++ web/src/components/header.tsx
@@ -5,10 +5,7 @@ import { CircleCheck, CircleDashed } from 'lucide-react';
 
 import DeckPathBreadcrumbs from '@/components/deck-path-breadcrumbs';
 import { SidebarTrigger, useSidebar } from '@/components/ui/sidebar';
-import {
-    getDeckDecksDeckIdGet,
-    getReviewSessionReviewSessionGet,
-} from '@/gen';
+import { getDeckDecksDeckIdGet, getReviewSessionReviewSessionGet } from '@/gen';
 
 export default function Header({ className }: React.ComponentProps<'div'>) {
     const { isMobile, openMobile } = useSidebar();
@@ -53,10 +50,12 @@ export default function Header({ className }: React.ComponentProps<'div'>) {
     const totalCards =
         remainingCards.length + failedCards.length + completedCards.length;
     const reviewedCards = completedCards.length + failedCards.length;
+    const completedPercentage = (completedCards.length / totalCards) * 100;
+    const failedPercentage = (failedCards.length / totalCards) * 100;
 
-    const showDeckBreadcrumbs = deckMatch && deck?.data;
-    const showReviewProgress = reviewMatch && totalCards > 0;
-    const hasContent = showDeckBreadcrumbs || showReviewProgress;
+    const onDecksPage = deckMatch && deck?.data;
+    const onReviewPage = reviewMatch && totalCards > 0;
+    const hasContent = onDecksPage || onReviewPage;
 
     // Hide on desktop if no content
     if (!hasContent && !isMobile) {
@@ -76,23 +75,49 @@ export default function Header({ className }: React.ComponentProps<'div'>) {
                 </div>
             )}
 
-            {showDeckBreadcrumbs && (
+            {onDecksPage && (
                 <DeckPathBreadcrumbs
                     path={deck.data.path}
                     showFullPath={true}
                 />
             )}
 
-            {showReviewProgress && (
-                <div className="text-muted-foreground flex items-center gap-2 text-sm">
-                    {totalCards == reviewedCards ? (
-                        <CircleCheck className="size-4" />
-                    ) : (
-                        <CircleDashed className="size-4" />
-                    )}
-                    <span>
-                        {reviewedCards} / {totalCards}
-                    </span>
+            {onReviewPage && (
+                <div className="flex w-full flex-row items-center justify-end gap-4 md:justify-center">
+                    {/*Progress bar*/}
+                    <div className="border-muted-foreground/20 flex h-3 w-[calc(100%-120px)] max-w-sm items-center overflow-hidden rounded-full border md:w-full">
+                        <div className="flex h-full flex-1 gap-[2px] overflow-hidden rounded-full">
+                            <div
+                                className={cn(
+                                    'bg-success h-full rounded-full',
+                                    completedCards.length == 0 ? 'hidden' : ''
+                                )}
+                                style={{
+                                    width: `${completedPercentage}%`,
+                                }}
+                            />
+                            <div
+                                className={cn(
+                                    'bg-error h-full rounded-full',
+                                    failedCards.length == 0 ? 'hidden' : ''
+                                )}
+                                style={{
+                                    width: `${failedPercentage}%`,
+                                }}
+                            />
+                        </div>
+                    </div>
+
+                    <div className="text-muted-foreground flex items-center gap-2 text-sm">
+                        {totalCards == reviewedCards ? (
+                            <CircleCheck className="size-4" />
+                        ) : (
+                            <CircleDashed className="size-4" />
+                        )}
+                        <span>
+                            {reviewedCards} / {totalCards}
+                        </span>
+                    </div>
                 </div>
             )}
         </header>
